Decided to use f_hat after all, after discussion with Fossen.
Thanks! But if Ad and Bd are updated in each time step (using the jacobians) isn't that how the EKF is supposed to work?

>>
There are two predictors in the EKF:
1) a linear predictor for the co-variance based on the linearized model (Ad, Bd, Cd, Dd)
2) a nonlinear state predictor based on the nonlinear state-space model. x[k+1] = f(x[k], u[k]) with  y[k] = h(x[k],u[k])

If you linearize the second equation you have a linear discrete-time Kalman filter. Extended Kalman filter means that you keep the state-space model nonlinear, see

https://en.wikipedia.org/wiki/Extended_Kalman_filter